Move Azure Key Vault logic to lib/az
