--- orchestrator/careplancontributor/keys/interface.go
@@ -1,9 +1,0 @@
-package keys
-
-import "crypto"
-
-type SigningKey interface {
-	crypto.Signer
-	SigningAlgorithm() string
-	KeyID() string
-}
