+++ orchestrator/lib/az/azkeyvault/encrypt_test.go
@@ -1,4 +1,4 @@
-package keys
+package azkeyvault
 
 import (
 	"net/http"
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestSigningKeyFromAzureKeyVault(t *testing.T) {
+func TestGetKey(t *testing.T) {
 	mux := http.NewServeMux()
 	httpServer := httptest.NewTLSServer(mux)
 	defer httpServer.Close()
@@ -22,7 +22,9 @@ func TestSigningKeyFromAzureKeyVault(t *testing.T) {
 	})
 
 	AzureHttpRequestDoer = httpServer.Client()
-	signingKey, err := signingKeyFromAzureKeyVault(httpServer.URL, "keyz", true)
+	client, err := NewClient(httpServer.URL, true)
+	require.NoError(t, err)
+	signingKey, err := GetKey(client, "keyz")
 	require.NoError(t, err)
 
 	assert.Equal(t, "https://keyszzz.vault.azure.net/keys/signingkey/5072fbaaa30849298e4b3c60384cdaac", signingKey.KeyID())
